[ci-app] Improve test environment metadata extraction
