+++ packages/dd-trace/src/plugins/util/test.js
@@ -1,3 +1,6 @@
+const { getGitMetadata, GIT_BRANCH, GIT_COMMIT_SHA, GIT_REPOSITORY_URL } = require('./git')
+const { getCIMetadata } = require('./ci')
+
 const TEST_FRAMEWORK = 'test.framework'
 const TEST_TYPE = 'test.type'
 const TEST_NAME = 'test.name'
@@ -9,5 +12,24 @@ module.exports = {
   TEST_TYPE,
   TEST_NAME,
   TEST_SUITE,
-  TEST_STATUS
+  TEST_STATUS,
+  getTestEnvironmentMetadata
+}
+
+function getTestEnvironmentMetadata (testFramework) {
+  // TODO: eventually these will come from the tracer (generally available)
+  const ciMetadata = getCIMetadata()
+  const {
+    [GIT_COMMIT_SHA]: commitSHA,
+    [GIT_BRANCH]: branch,
+    [GIT_REPOSITORY_URL]: repositoryUrl
+  } = ciMetadata
+
+  const gitMetadata = getGitMetadata({ commitSHA, branch, repositoryUrl })
+
+  return {
+    [TEST_FRAMEWORK]: testFramework,
+    ...gitMetadata,
+    ...ciMetadata
+  }
 }
